+++ resdk/resources/data.py
@@ -1,6 +1,7 @@
 """Data resource."""
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+import json
 import logging
 
 import requests
@@ -156,8 +157,57 @@ class Data(BaseResource):
             self._sample = None if len(self._sample) == 0 else self._sample[0]
         return self._sample
 
+    def _files_dirs(self, field_type, file_name=None, field_name=None):
+        """Get list of downloadable fields."""
+        download_list = []
+
+        def put_in_download_list(elm, fname):
+            """Append only files od dirs with equal name."""
+            if field_type in elm:
+                if file_name is None or file_name == elm[field_type]:
+                    download_list.append(elm[field_type])
+            else:
+                raise KeyError("Item {} does not contain '{}' key.".format(fname, field_type))
+
+        if field_name and not field_name.startswith('output.'):
+            field_name = 'output.{}'.format(field_name)
+
+        for ann_field_name, ann in self.annotation.items():
+            if (ann_field_name.startswith('output') and
+                    (field_name is None or field_name == ann_field_name) and
+                    ann['value'] is not None):
+                if ann['type'].startswith('basic:{}:'.format(field_type)):
+                    put_in_download_list(ann['value'], ann_field_name)
+                elif ann['type'].startswith('list:basic:{}:'.format(field_type)):
+                    for element in ann['value']:
+                        put_in_download_list(element, ann_field_name)
+
+        return download_list
+
+    def _get_dir_files(self, dir_name):
+        files_list, dir_list = [], []
+
+        dir_url = urljoin(self.resolwe.url, 'data/{}/{}'.format(self.id, dir_name))
+        if not dir_url.endswith('/'):
+            dir_url += '/'
+        response = requests.get(dir_url, auth=self.resolwe.auth)
+        response = json.loads(response.content.decode('utf-8'))
+
+        for obj in response:
+            obj_path = '{}/{}'.format(dir_name, obj['name'])
+            if obj['type'] == 'directory':
+                dir_list.append(obj_path)
+            else:
+                files_list.append(obj_path)
+
+        if dir_list:
+            for new_dir in dir_list:
+                files_list.extend(self._get_dir_files(new_dir))
+
+        return files_list
+
     def files(self, file_name=None, field_name=None):
-        """Get list of downloadable fields.
+        """Get list of downloadable file fields.
 
         Filter files by file name or output field.
 
@@ -168,47 +218,34 @@ class Data(BaseResource):
         :rtype: List of tuples (data_id, file_name, field_name, process_type)
 
         """
-        download_list = []
+        if not self.id:
+            raise ValueError('Instance must be saved before using `files` method.')
 
-        def put_in_download_list(elm, fname):
-            """Append only files with equal name."""
-            if 'file' in elm:
-                if file_name is None or file_name == elm['file']:
-                    download_list.append(elm['file'])
-            else:
-                raise KeyError("Item {} does not contain 'file' key.".format(fname))
+        file_list = self._files_dirs('file', file_name, field_name)
 
-        if field_name and not field_name.startswith('output.'):
-            field_name = 'output.{}'.format(field_name)
+        for dir_name in self._files_dirs('dir', file_name, field_name):
+            print('>>', self._get_dir_files)
+            file_list.extend(self._get_dir_files(dir_name))
 
-        for ann_field_name, ann in self.annotation.items():
-            if (ann_field_name.startswith('output') and
-                    (field_name is None or field_name == ann_field_name) and
-                    ann['value'] is not None):
-                if ann['type'].startswith('basic:file:'):
-                    put_in_download_list(ann['value'], ann_field_name)
-                elif ann['type'].startswith('list:basic:file:'):
-                    for element in ann['value']:
-                        put_in_download_list(element, ann_field_name)
-
-        return download_list
+        return file_list
 
     def download(self, file_name=None, field_name=None, download_dir=None):
-        """Download Data object's files.
+        """Download Data object's files and directories.
 
-        Download files from the Resolwe server to the download
-        directory (defaults to the current working directory).
+        Download files and directoriesfrom the Resolwe server to the
+        download directory (defaults to the current working directory).
 
-        :param file_name: name of file
+        :param file_name: name of file or directory
         :type file_name: string
-        :param field_name: file field name
+        :param field_name: file or directory field name
         :type field_name: string
         :param download_dir: download path
         :type download_dir: string
         :rtype: None
 
-        Data objects can contain multiple files. All are downloaded by
-        default, but may be filtered by name or output field:
+        Data objects can contain multiple files and directories. All are
+        downloaded by default, but may be filtered by name or output
+        field:
 
         * re.data.get(42).download(file_name='alignment7.bam')
         * re.data.get(42).download(field_name='bam')